Move chat parsing to shared file
