+++ src/providers/azureopenai.ts
@@ -207,35 +207,8 @@ export class AzureOpenAiChatCompletionProvider extends AzureOpenAiGenericProvide
       throw new Error('Azure OpenAI API host must be set');
     }
 
-    let messages: { role: string; content: string; name?: string }[];
-    const trimmedPrompt = prompt.trim();
-    if (trimmedPrompt.startsWith('- role:')) {
-      try {
-        // Try YAML
-        messages = yaml.load(prompt) as { role: string; content: string }[];
-      } catch (err) {
-        throw new Error(
-          `Azure OpenAI Chat Completion prompt is not a valid YAML string: ${err}\n\n${prompt}`,
-        );
-      }
-    } else {
-      try {
-        // Try JSON
-        messages = JSON.parse(prompt) as { role: string; content: string }[];
-      } catch (err) {
-        const trimmedPrompt = prompt.trim();
-        if (
-          process.env.PROMPTFOO_REQUIRE_JSON_PROMPTS ||
-          trimmedPrompt.startsWith('{') ||
-          trimmedPrompt.startsWith('[')
-        ) {
-          throw new Error(
-            `Azure OpenAI Chat Completion prompt is not a valid JSON string: ${err}\n\n${prompt}`,
-          );
-        }
-        messages = [{ role: 'user', content: prompt }];
-      }
-    }
+    import { parsePrompt } from './shared';
+    const messages = parsePrompt(prompt);
 
     const body = {
       model: this.deploymentName,
