+++ src/providers/openai.ts
@@ -227,36 +227,8 @@ export class OpenAiChatCompletionProvider extends OpenAiGenericProvider {
       );
     }
 
-    let messages: { role: string; content: string; name?: string }[];
-    const trimmedPrompt = prompt.trim();
-    if (trimmedPrompt.startsWith('- role:')) {
-      try {
-        // Try YAML
-        messages = yaml.load(prompt) as { role: string; content: string }[];
-      } catch (err) {
-        throw new Error(
-          `OpenAI Chat Completion prompt is not a valid YAML string: ${err}\n\n${prompt}`,
-        );
-      }
-    } else {
-      try {
-        // Try JSON
-        messages = JSON.parse(prompt) as { role: string; content: string }[];
-      } catch (err) {
-        if (
-          process.env.PROMPTFOO_REQUIRE_JSON_PROMPTS ||
-          trimmedPrompt.startsWith('{') ||
-          trimmedPrompt.startsWith('[')
-        ) {
-          throw new Error(
-            `OpenAI Chat Completion prompt is not a valid JSON string: ${err}\n\n${prompt}`,
-          );
-        }
-
-        // Fall back to wrapping the prompt in a user message
-        messages = [{ role: 'user', content: prompt }];
-      }
-    }
+    import { parsePrompt } from './shared';
+    const messages = parsePrompt(prompt);
 
     const body = {
       model: this.modelName,
